pyproject.toml
[tool.poetry]
name = "transactions-processing"
version = "0.1.0"
description = ""
authors = ["Danil Sidoruk <danilsidoruknv@yandex.ru>"]
readme = "README.md"
packages = [{include = "transactions_processing"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.6"


[tool.poetry.group.dev.dependencies]
pylint = "^2.17.4"
isort = "^5.12.0"
black = "^23.3.0"
pre-commit = "^3.3.2"

[tool.black]
line-length = 120
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 120
multi_line_output = 3
include_trailing_comma = true
use_parentheses = true
ensure_newline_before_comments = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
